requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
psycopg==3.2.3
psycopg2-binary==2.9.10
requests==2.32.3
sorl-thumbnail==12.11.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
